refactor: split filtering steps into dedicated files
